[UIMA-2369] don't overwrite local repo if file being downloaded already exists.
git-svn-id: https://svn.apache.org/repos/asf/uima/build/trunk@1301120 13f79535-47bb-0310-9956-ffa450edef68

+++ uima-build-helper-maven-plugin/src/main/java/org/apache/uima/buildhelper/CopyFromApacheDist.java
@@ -119,8 +119,12 @@ public class CopyFromApacheDist extends AbstractMojo {
     }
   
     FileOutputStream os = null;
-    try {
-      targetFile.getParentFile().mkdirs();
+    targetFile.getParentFile().mkdirs();
+    if (targetFile.exists()) {
+      System.out.format(" *** Surprise, file %s exists on 2nd check%n", targetInLocalFileSystem);
+      return;
+    }
+    try {      
       os = new FileOutputStream(targetFile);
     } catch (FileNotFoundException e) {
       throw new MojoExecutionException("While creating local file in location " + targetFile.getAbsolutePath(), e);
